fabric: Remove unfinished argument types, add some javadoc, fix checkstyle issues
