+++ .checkstyle/checkstyle.xml
@@ -100,6 +100,7 @@
     <module name="LineLength">
         <property name="fileExtensions" value="java"/>
         <property name="max" value="140"/>
+        <property name="ignorePattern" value="^ *\* "/> <!-- Ignore long JavaDoc lines (i.e. a link) -->
     </module>
 
     <!-- Checks for whitespace                               -->
